Create a custom kable table.

kkable builds off of knitr::kable() as well as kableExtra::kable_styling() to create a customized table for use in R Markdown.

kkable(
  data,
  format = "html",
  col_names = "auto",
  alignment = "auto",
  proper_cols = NULL,
  currency_cols = NULL,
  number_cols = NULL,
  decimal_cols = NULL,
  decimal_digits = 2,
  percent_cols = NULL,
  percent_digits = 2,
  format_header = TRUE,
  add_totals = FALSE,
  format_totals = add_totals,
  digits = getOption("digits"),
  row.names = FALSE,
  caption = NULL,
  label = NULL,
  format.args = list(),
  escape = TRUE,
  bootstrap_options = c("striped", "bordered", "responsive"),
  full_width = NULL,
  position = "center",
  font_size = NULL,
  row_label_position = "l",
  repeat_header_text = "\\textit{(continued)}",
  repeat_header_method = c("append", "replace"),
  repeat_header_continued = FALSE,
  stripe_color = "gray!6",
  stripe_index = NULL,
  fixed_thead = FALSE,
  ...
)

Arguments

data
An R object, typically a matrix or data frame.

col_names
column names to display - defaults to "auto" which attempts to formalize the column names for you. Alternatively, you can insert a character vector of column names.

alignment
mimics knitr::kable() align argument but with an "auto" option.

currency_cols
specify currency columns which will be formatted via formattable::currency().

number_cols
specify numeric columns which will be formatted via formattable::comma().

decimal_cols
specify decimal columns (i.e. numeric columns where you want to show decimal places).

decimal_digits
number of digits to show in decimal cols.

percent_cols
specify percent columns which will be formatted via formattable::percent().

percent_digits
number of digits to display for percent cols.

format_header
logical - if TRUE will style the tables header.

add_totals
Logical: whether to append a totals row to the output. If TRUE uses janitor::adorn_totals(fill = " ") to apply the total row. See janitor::adorn_totals() for more information.

format_totals
logical - if TRUE will style the last row as a total row.

bootstrap_options
A character vector for bootstrap table options. Please see package vignette or visit the w3schools' Bootstrap Page for more information. Possible options include basic, striped, bordered, hover, condensed and responsive.

full_width
A TRUE or FALSE variable controlling whether the HTML table should have 100% width. Since HTML and pdf have different flavors on the preferable format for full_width. If not specified, a HTML table will have full width by default but this option will be set to FALSE for a LaTeX table

position
A character string determining how to position the table on a page. Possible values include left, center, right, float_left and float_right. Please see the package doc site for demonstrations. For a LaTeX table, if float_* is selected, LaTeX package wrapfig will be imported.

font_size
A numeric input for table font size

row_label_position
A character string determining the justification of the row labels in a table. Possible values inclued l for left, c for center, and r for right. The default value is l for left justifcation.

repeat_header_text
LaTeX option. A text string you want to append on or replace the caption.

repeat_header_method
LaTeX option, can either be append(default) or replace

repeat_header_continued
T/F or a text string. Whether or not to put a continued mark on the second page of longtable. If you put in text, we will use this text as the "continued" mark.

stripe_color
LaTeX option allowing users to pick a different color for their strip lines. This option is not available in HTML

stripe_index
LaTeX option allowing users to customize which rows should have stripe color.

fixed_thead
HTML table option so table header row is fixed at top. Values can be either T/F or list(enabled = T/F, background = "anycolor").

...
Arguments passed on to knitr::kable

x

For kable(), x is an R object, which is typically a matrix or data frame. For kables(), a list with each element being a returned value from kable().

format

A character string. Possible values are latex, html, pipe (Pandoc's pipe tables), simple (Pandoc's simple tables), and rst. The value of this argument will be automatically determined if the function is called within a knitr document. The format value can also be set in the global option knitr.table.format. If format is a function, it must return a character string.

digits

Maximum number of digits for numeric columns, passed to round(). This can also be a vector of length ncol(x), to set the number of digits for individual columns.

row.names

Logical: whether to include row names. By default, row names are included if rownames(x) is neither NULL nor identical to 1:nrow(x).

col.names

A character vector of column names to be used in the table.

align

Column alignment: a character vector consisting of 'l' (left), 'c' (center) and/or 'r' (right). By default or if align = NULL, numeric columns are right-aligned, and other columns are left-aligned. If length(align) == 1L, the string will be expanded to a vector of individual letters, e.g. 'clc' becomes c('c', 'l', 'c'), unless the output format is LaTeX.

caption

The table caption.

label

The table reference label. By default, the label is obtained from knitr::opts_current$get('label').

format.args

A list of arguments to be passed to format() to format table values, e.g. list(big.mark = ',').

escape

Boolean; whether to escape special characters when producing HTML or LaTeX tables. When escape = FALSE, you have to make sure that special characters will not trigger syntax errors in LaTeX or HTML.

Value

a customized table for display of class ""kableExtra" and/or "knitr_kable".

Note

The tables for format = 'markdown' also work for Pandoc when the pipe_tables extension is enabled (this is the default behavior for Pandoc >= 1.10).

References

See https://github.com/yihui/knitr-examples/blob/master/091-knitr-table.Rnw for some examples in LaTeX, but they also apply to other document formats.

Examples

library(propalloc)

initial_costs <- propalloc::renewal_costs

propalloc:::kkable(
 initial_costs,
 caption = "Initial Costs Table",
 proper_cols = c("cost_type", "description"),
 currency_cols = c("prior", "current"),
 digits = 0,
 format_header = TRUE,
 add_totals = TRUE,
 format_totals = TRUE
)

#> <table class="table table-striped table-bordered table-responsive" style="margin-left: auto; margin-right: auto;">
#> <caption>Initial Costs Table</caption>
#>  <thead>
#>   <tr>
#>    <th style="text-align:center;font-weight: bold;text-align: center;"> Cost Type </th>
#>    <th style="text-align:center;font-weight: bold;text-align: center;"> Description </th>
#>    <th style="text-align:right;font-weight: bold;text-align: center;"> Prior </th>
#>    <th style="text-align:right;font-weight: bold;text-align: center;"> Current </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:center;"> Premium </td>
#>    <td style="text-align:center;"> All Risk </td>
#>    <td style="text-align:right;"> $63,301,745 </td>
#>    <td style="text-align:right;"> $65,200,798 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:center;"> Premium </td>
#>    <td style="text-align:center;"> Terrorism </td>
#>    <td style="text-align:right;"> $3,909,519 </td>
#>    <td style="text-align:right;"> $4,026,804 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:center;"> Expense </td>
#>    <td style="text-align:center;"> Taxes </td>
#>    <td style="text-align:right;"> $602,831 </td>
#>    <td style="text-align:right;"> $620,916 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:center;"> Expense </td>
#>    <td style="text-align:center;"> Fees </td>
#>    <td style="text-align:right;"> $4,645,560 </td>
#>    <td style="text-align:right;"> $4,784,927 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:center;"> Expense </td>
#>    <td style="text-align:center;"> Program </td>
#>    <td style="text-align:right;"> $1,435,010 </td>
#>    <td style="text-align:right;"> $1,478,060 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:center;font-weight: bold;background-color: lightgrey !important;"> Total </td>
#>    <td style="text-align:center;font-weight: bold;background-color: lightgrey !important;">  </td>
#>    <td style="text-align:right;font-weight: bold;background-color: lightgrey !important;"> $73,894,665 </td>
#>    <td style="text-align:right;font-weight: bold;background-color: lightgrey !important;"> $76,111,505 </td>
#>   </tr>
#> </tbody>
#> </table>